+++ src/Util/DateRange.php
@@ -11,10 +11,31 @@ final class DateRange
     private ?Chronos $startDate;
     private ?Chronos $endDate;
 
+    /** @deprecated Use named constructors */
     public function __construct(?Chronos $startDate = null, ?Chronos $endDate = null)
     {
         $this->startDate = $startDate;
         $this->endDate = $endDate;
+    }
+
+    public static function emptyInstance(): self
+    {
+        return new self();
+    }
+
+    public static function withStartDate(Chronos $startDate): self
+    {
+        return new self($startDate);
+    }
+
+    public static function withEndDate(Chronos $endDate): self
+    {
+        return new self(null, $endDate);
+    }
+
+    public static function withStartAndEndDate(Chronos $startDate, Chronos $endDate): self
+    {
+        return new self($startDate, $endDate);
     }
 
     public function getStartDate(): ?Chronos